separate the debugger logic from the controller
