+++ lib/api/controllers/debugController.ts
@@ -21,40 +21,12 @@
 
 import { KuzzleRequest } from "../request";
 import { NativeController } from "./baseController";
-import Inspector from "inspector";
 import * as kerror from "../../kerror";
-import { DebugModule } from "../../types/DebugModule";
-import { JSONObject } from "kuzzle-sdk";
-import get from "lodash/get";
-
-const DEBUGGER_EVENT = "kuzzle-debugger-event";
-
-type DebugModuleMethod = (params: JSONObject) => any;
 
 /**
  * @class DebugController
  */
 export class DebugController extends NativeController {
-  private inspector: Inspector.Session;
-
-  private debuggerStatus = false;
-
-  /**
-   * Map<eventName, Set<connectionId>>
-   */
-  private events = new Map<string, Set<string>>();
-
-  /**
-   * Map of functions from the DebugModules
-   */
-  private kuzzlePostMethods = new Map<string, DebugModuleMethod>();
-
-  /**
-   * List of DebugModule for DebugController
-   * Used to add new methods and events to the protocol
-   */
-  private modules: DebugModule[] = [];
-
   constructor() {
     super([
       "nodeVersion",
@@ -66,49 +38,9 @@ export class DebugController extends NativeController {
     ]);
   }
 
-  async init() {
-    super.init();
-
-    this.inspector = new Inspector.Session();
-
-    // Remove connection id from the list of listeners for each event
-    global.kuzzle.on("connection:remove", (connectionId) => {
-      if (!this.debuggerStatus) {
-        return;
-      }
-
-      for (const listener of this.events.values()) {
-        listener.delete(connectionId);
-      }
-    });
-
-    this.inspector.on("inspectorNotification", async (payload) => {
-      if (!this.debuggerStatus) {
-        return;
-      }
-
-      await this.notifyGlobalListeners(payload.method, payload);
-
-      const listeners = this.events.get(payload.method);
-      if (!listeners) {
-        return;
-      }
-
-      const promises = [];
-      for (const connectionId of listeners) {
-        promises.push(
-          this.notifyConnection(connectionId, DEBUGGER_EVENT, {
-            event: payload.method,
-            result: payload,
-          })
-        );
-      }
-
-      // No need to catch, notify is already try-catched
-      await Promise.all(promises);
-    });
-  }
-
+  /**
+   * Return the node version of the current Kuzzle instance
+   */
   async nodeVersion() {
     return process.version;
   }
@@ -117,78 +49,14 @@ export class DebugController extends NativeController {
    * Connect the debugger
    */
   async enable() {
-    if (this.debuggerStatus) {
-      return;
-    }
-
-    this.inspector.connect();
-    this.debuggerStatus = true;
-
-    for (const module of this.modules) {
-      await module.init(this.inspector);
-
-      for (const methodName of module.methods) {
-        if (!module[methodName]) {
-          throw new Error(
-            `Missing implementation of method "${methodName}" inside DebugModule "${module.name}"`
-          );
-        }
-        this.kuzzlePostMethods.set(
-          `Kuzzle.${module.name}.${methodName}`,
-          module[methodName].bind(module)
-        );
-      }
-
-      for (const eventName of module.events) {
-        module.on(eventName, async (payload) => {
-          if (!this.debuggerStatus) {
-            return;
-          }
-
-          const event = `Kuzzle.${module.name}.${eventName}`;
-          await this.notifyGlobalListeners(event, payload);
-
-          const listeners = this.events.get(event);
-          if (!listeners) {
-            return;
-          }
-
-          const promises = [];
-          for (const connectionId of listeners) {
-            promises.push(
-              this.notifyConnection(connectionId, DEBUGGER_EVENT, {
-                event,
-                result: payload,
-              })
-            );
-          }
-
-          // No need to catch, notify is already try-catched
-          await Promise.all(promises);
-        });
-      }
-    }
+    global.kuzzle.ask("core:debugger:enable");
   }
 
   /**
    * Disconnect the debugger and clears all the events listeners
    */
   async disable() {
-    if (!this.debuggerStatus) {
-      return;
-    }
-
-    for (const module of this.modules) {
-      for (const eventName of module.events) {
-        module.removeAllListeners(eventName);
-      }
-      await module.cleanup();
-    }
-
-    this.inspector.disconnect();
-    this.debuggerStatus = false;
-    this.events.clear();
-    this.kuzzlePostMethods.clear();
+    global.kuzzle.ask("core:debugger:disable");
   }
 
   /**
@@ -196,31 +64,10 @@ export class DebugController extends NativeController {
    * See: https://chromedevtools.github.io/devtools-protocol/v8/
    */
   async post(request: KuzzleRequest) {
-    if (!this.debuggerStatus) {
-      throw kerror.get("core", "debugger", "not_enabled");
-    }
-
     const method = request.getBodyString("method");
     const params = request.getBodyObject("params", {});
 
-    if (method.startsWith("Kuzzle.")) {
-      const debugModuleMethod = this.kuzzlePostMethods.get(method);
-
-      if (debugModuleMethod) {
-        return debugModuleMethod(params);
-      }
-      throw kerror.get("core", "debugger", "method_not_found", method);
-    }
-
-    if (!get(global.kuzzle.config, "security.debug.native_debug_protocol")) {
-      throw kerror.get(
-        "core",
-        "debugger",
-        "native_debug_protocol_usage_denied"
-      );
-    }
-
-    return this.inspectorPost(method, params);
+    return global.kuzzle.ask('core:debugger:post', method, params);
   }
 
   /**
@@ -238,19 +85,9 @@ export class DebugController extends NativeController {
       );
     }
 
-    if (!this.debuggerStatus) {
-      throw kerror.get("core", "debugger", "not_enabled");
-    }
-
     const event = request.getBodyString("event");
 
-    let listeners = this.events.get(event);
-    if (!listeners) {
-      listeners = new Set();
-      this.events.set(event, listeners);
-    }
-
-    listeners.add(request.context.connection.id);
+    global.kuzzle.ask("core:debugger:addListener", event, request.context.connection.id);
   }
 
   /**
@@ -267,85 +104,8 @@ export class DebugController extends NativeController {
       );
     }
 
-    if (!this.debuggerStatus) {
-      throw kerror.get("core", "debugger", "not_enabled");
-    }
-
     const event = request.getBodyString("event");
-
-    const listeners = this.events.get(event);
-
-    if (listeners) {
-      listeners.delete(request.context.connection.id);
-    }
-  }
-
-  /**
-   * Execute a method using the Chrome Debug Protocol
-   * @param method Chrome Debug Protocol method to execute
-   * @param params
-   * @returns
-   */
-  private async inspectorPost(
-    method: string,
-    params: JSONObject
-  ): Promise<JSONObject> {
-    if (!this.debuggerStatus) {
-      throw kerror.get("core", "debugger", "not_enabled");
-    }
-
-    let resolve;
-
-    const promise = new Promise((res) => {
-      resolve = res;
-    });
-
-    this.inspector.post(method, params, (err, res) => {
-      if (err) {
-        resolve({
-          error: JSON.stringify(Object.getOwnPropertyDescriptors(err)),
-        });
-      } else {
-        resolve(res);
-      }
-    });
-
-    return promise;
-  }
-
-  /**
-   * Sends a direct notification to a websocket connection without having to listen to a specific room
-   */
-  private async notifyConnection(
-    connectionId: string,
-    event: string,
-    payload: JSONObject
-  ) {
-    global.kuzzle.entryPoint._notify({
-      channels: [event],
-      connectionId,
-      payload,
-    });
-  }
-
-  private async notifyGlobalListeners(event: string, payload: JSONObject) {
-    const listeners = this.events.get("*");
-
-    if (!listeners) {
-      return;
-    }
-
-    const promises = [];
-    for (const connectionId of listeners) {
-      promises.push(
-        this.notifyConnection(connectionId, DEBUGGER_EVENT, {
-          event,
-          result: payload,
-        })
-      );
-    }
-
-    // No need to catch, notify is already try-catched
-    await Promise.all(promises);
+    
+    global.kuzzle.ask("core:debugger:removeListener", event, request.context.connection.id);
   }
 }
